Cast float to double in FloatingPointArrayConverters
Fix #2480

+++ Source/HelixToolkit.Wpf/ExtensionMethods/FloatingPointArrayConverters.cs
@@ -72,7 +72,7 @@ internal static class FloatingPointArrayConverters
 
                 for (int i = length * floatCount; i < arrayCount; i++)
                 {
-                    *doubleData = *floatData;
+                    *doubleData = (double)*floatData;
                     floatData++;
                     doubleData++;
                 }
